Added initializer base
