+++ main.cpp
@@ -15,24 +15,13 @@ using namespace std;
 #include <GLFW/glfw3.h>
 #include <GL/gl.h>
 
+#include "Initializers/BaseInitializer.hpp"
+
 static GLFWwindow *window = nullptr;
 static int frames = 0;
 
 static Config config;
 
-struct Textures {
-    GLuint mainTextureID;
-    GLuint particleCountTextureID;
-    GLuint particlePositionTextureID;
-    GLuint particleDestinationTextureID;
-    GLuint particleMassTextureID;
-    GLuint colorTextureRID;
-    GLuint colorTextureGID;
-    GLuint colorTextureBID;
-    GLuint imageTextureID;
-    int imageWidth;
-    int imageHeight;
-};
 
 static void printHelp(const char *arg0);
 static void printConfig();
@@ -48,8 +37,8 @@ int main(int argc, char **argv)
     handleArguments(argc, argv);
     initializeGLFW();
 
-    Textures textures{};
-    initializeTextures(textures);
+    BaseInitializer* initializer = new BaseInitializer();
+    Textures textures = initializer->initializeTextures(config);
 
 
 	// Create the render program
@@ -181,158 +170,6 @@ void generateViewPane(GLuint renderProgramID) {// Two triangles
     glEnableVertexAttribArray(posPtr);
 }
 
-void initializeTextures(Textures &textures) {
-    // Create the main texture
-    {
-        auto sz = config.width * config.height;
-        float *data = new float[sz * 4];
-        for (int i = 0; i < sz; ++i) {
-            float c = float(i) / sz;
-            data[i * 4 + 0] = c;
-            data[i * 4 + 1] = c;
-            data[i * 4 + 2] = c;
-            data[i * 4 + 3] = 1.0;
-        }
-
-        glGenTextures(1, &textures.mainTextureID);
-        glBindTexture(GL_TEXTURE_2D, textures.mainTextureID);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
-        glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, config.width, config.height, 0, GL_RGBA, GL_FLOAT, data);
-        delete[] data;
-    }
-
-    // Create a integer texture, where every pixel's value is the number of particles
-    // that should be positioned on the corresponding pixel on the main texture/screen
-    {
-        auto sz = config.width * config.height * 4;
-        GLuint *data = new GLuint[sz];
-        for (int i = 0; i < sz; ++i)
-            data[i] = 2;
-        glGenTextures(1, &textures.particleCountTextureID);
-        glBindTexture(GL_TEXTURE_2D, textures.particleCountTextureID);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
-        glTexImage2D(GL_TEXTURE_2D, 0, GL_R32UI, config.width, config.height, 0, GL_RED_INTEGER, GL_UNSIGNED_INT, data);
-        delete[] data;
-    }
-
-    // Create a texture in which the particle coordinates will be stored
-    {
-        int sizeX = config.particleCountX;
-        int sizeY = config.particleCountY;
-
-        auto sz = sizeX * sizeY;
-        float *data = new float[sz * 4];
-        for (int i = 0; i < sz; ++i) {
-            data[i * 4 + 0] = config.width * (rand() / float(RAND_MAX));
-            data[i * 4 + 1] = config.height * (rand() / float(RAND_MAX));
-            data[i * 4 + 2] = 0.0;
-            data[i * 4 + 3] = 0.0;
-        }
-
-        glGenTextures(1, &textures.particlePositionTextureID);
-        glBindTexture(GL_TEXTURE_2D, textures.particlePositionTextureID);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
-        glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, sizeX, sizeY, 0, GL_RGBA, GL_FLOAT, data);
-        delete[] data;
-    }
-
-    // Create a texture in which the particle destinations are
-    {
-        int sizeX = config.particleCountX;
-        int sizeY = config.particleCountY;
-
-        auto sz = sizeX * sizeY;
-        float *data = new float[sz * 4];
-        for (int i = 0; i < sz; ++i) {
-            float positionX = (rand() / float(RAND_MAX)) * config.width;
-            float positionY = (rand() / float(RAND_MAX)) * config.height;
-//            float radius = sqrt(rand() / float(RAND_MAX));
-//            data[i * 4 + 0] = 150 + 50.0f * cos(angle) * radius;
-//            data[i * 4 + 1] = 150 + 50.0f * sin(angle) * radius;
-            data[i * 4 + 0] = positionX;
-            data[i * 4 + 1] = positionY;
-            data[i * 4 + 2] = 0.0;
-            data[i * 4 + 3] = 0.0;
-        }
-
-        glGenTextures(1, &textures.particleDestinationTextureID);
-        glBindTexture(GL_TEXTURE_2D, textures.particleDestinationTextureID);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
-        glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, sizeX, sizeY, 0, GL_RGBA, GL_FLOAT, data);
-        delete[] data;
-    }
-
-    // Create a texture in which the particle masses will be stored
-    {
-        int sizeX = config.particleCountX;
-        int sizeY = config.particleCountY;
-
-        auto sz = sizeX * sizeY;
-        float *data = new float[sz];
-        for (int i = 0; i < sz; ++i) {
-            data[i] = config.massMin + (config.massMax - config.massMin) * (float(rand()) / RAND_MAX);
-        }
-
-        glGenTextures(1, &textures.particleMassTextureID);
-        glBindTexture(GL_TEXTURE_2D, textures.particleMassTextureID);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
-        glTexImage2D(GL_TEXTURE_2D, 0, GL_R32F, sizeX, sizeY, 0, GL_RED, GL_FLOAT, data);
-        delete[] data;
-    }
-
-    // Create a texture in which the output color will be stored
-    {
-        auto sz = config.width * config.height;
-        GLuint  *data = new GLuint[sz];
-        for (int i = 0; i < sz; ++i) {
-            data[i] = 1;
-        }
-
-        glGenTextures(1, &textures.colorTextureRID);
-        glBindTexture(GL_TEXTURE_2D, textures.colorTextureRID);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
-        glTexImage2D(GL_TEXTURE_2D, 0, GL_R32UI, config.width, config.height, 0, GL_RED_INTEGER, GL_UNSIGNED_INT, data);
-        glGenTextures(1, &textures.colorTextureGID);
-        glBindTexture(GL_TEXTURE_2D, textures.colorTextureGID);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
-        glTexImage2D(GL_TEXTURE_2D, 0, GL_R32UI, config.width, config.height, 0, GL_RED_INTEGER, GL_UNSIGNED_INT, data);
-        glGenTextures(1, &textures.colorTextureBID);
-        glBindTexture(GL_TEXTURE_2D, textures.colorTextureBID);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
-        glTexImage2D(GL_TEXTURE_2D, 0, GL_R32UI, config.width, config.height, 0, GL_RED_INTEGER, GL_UNSIGNED_INT, data);
-        delete[] data;
-    }
-
-    {
-        const char* imagePath = "images/icaka.png";
-        int nrChannels;
-        unsigned char *data = stbi_load(imagePath, &textures.imageWidth, &textures.imageHeight, &nrChannels, 0);
-        int sz = textures.imageWidth*textures.imageHeight*4;
-        float *fdata = new float[sz];
-        for (int i = 0; i < sz; ++i) {
-            fdata[i] = data[i];
-        }
-
-        if (!data) {
-            cerr << "Failed loading image: " << imagePath << endl;
-        }
-
-        glGenTextures(1, &textures.imageTextureID);
-        glBindTexture(GL_TEXTURE_2D, textures.imageTextureID);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
-        glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, textures.imageWidth, textures.imageHeight, 0, GL_RGBA, GL_FLOAT, fdata);
-        stbi_image_free(data);
-    }
-}
 
 void printHelp(const char *arg0) {
     cout << "Usage: " << arg0 << " [OPTIONS]" << endl << endl
